# api/swagger.json
{
   "openapi": "3.0.0",
   "info": {
      "description": "This is api design for binar full-stack assessment test",
      "version": "1.0.0",
      "title": "Swagger Movie DB API",
      "termsOfService": "http://swagger.io/terms/",
      "contact": {
         "email": "apiteam@swagger.io"
      },
      "license": {
         "name": "Apache 2.0",
         "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
   },
...
   "paths": {
...
      "/genres/{genreId}": {
         "get": {
            "tags": [
               "genres"
            ],
            "summary": "Get detail genre data in particular genreId",
            "operationId": "getGenreByID",
            "parameters": [
               {
                  "name": "genreId",
                  "in": "path",
                  "description": "ID of genre to retrieve specific genre data",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "format": "int32"
                  }
               },
               {
                  "in": "header",
                  "name": "Authorization",
                  "description": "Authorization bearer token",
                  "required": true,
                  "schema": {
                     "type": "string",
                     "format": "JWT"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Genre"
                        }
                     }
                  }
               },
               "400": {
                  "description": "Invalid ID supplied"
               },
               "401": {
                  "description": "Unauthorized access"
               },
               "404": {
                  "description": "Genre not found"
               }
            },
            "security": [
               {
                  "bearerAuth": []
               }
            ]
         },
         "put": {
            "tags": [
               "genres"
            ],
            "summary": "Update specific genre data according to genre ID",
            "operationId": "updateGenreByID",
            "parameters": [
               {
                  "name": "genreId",
                  "in": "path",
                  "description": "ID of genre to retrieve specific genre data",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "format": "int32"
                  }
               },
               {
                  "name": "Authorization",
                  "in": "header",
                  "required": true,
                  "schema": {
                     "type": "string",
                     "format": "JWT"
                  }
               }
            ],
            "requestBody": {
               "description": "Updated genre object",
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "anyOf": [
                           {
                              "$ref": "#/components/schemas/Genre"
                           }
                        ]
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "successful operation",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Genre"
                        }
                     }
                  }
               },
               "400": {
                  "description": "Invalid ID supplied"
               },
               "401": {
                  "description": "Unauthorized access"
               },
               "404": {
                  "description": "Genre not found"
               }
            },
            "security": [
               {
                  "bearerAuth": []
               }
            ]
         },
         "delete": {
            "tags": [
               "genres"
            ],
            "summary": "Delete genre data based on defined ID",
            "description": "",
            "operationId": "deleteGenre",
            "parameters": [
               {
                  "in": "header",
                  "name": "Authorization",
                  "required": true,
                  "schema": {
                     "type": "string",
                     "format": "JWT"
                  }
               },
               {
                  "name": "genreId",
                  "in": "path",
                  "description": "ID of genre to retrieve specific genre data",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "format": "int32"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "OK",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "message": {
                                 "type": "string"
                              }
                           }
                        }
                     }
                  }
               },
               "400": {
                  "description": "Invalid ID supplied"
               },
               "401": {
                  "description": "Unauthorized access"
               },
               "404": {
                  "description": "Genre not found"
               }
            },
            "security": [
               {
                  "bearerAuth": []
               }
            ]
         }
      },
...
      "/status": {
         "post": {
            "tags": [
               "status"
            ],
            "summary": "Add a new status",
            "description": "Add a new movie status. Movie status is divided into rumored, planned, in production, post production, released, and cancelled",
            "operationId": "createMovieStatus",
            "parameters": [
               {
                  "name": "Authorization",
                  "in": "header",
                  "required": true,
                  "schema": {
                     "type": "string",
                     "format": "JWT"
                  }
               }
            ],
            "requestBody": {
               "description": "Status object that needs to be added.",
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/Status"
                     }
                  }
               }
            },
            "responses": {
               "201": {
                  "description": "Created",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Status"
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized access"
               },
               "405": {
                  "description": "Invalid input"
               }
            },
            "security": [
               {
                  "bearerAuth": []
               }
            ]
         },
         "get": {
            "tags": [
               "status"
            ],
            "summary": "Get all existing movie status",
            "description": "",
            "operationId": "getMovieStatus",
            "parameters": [
               {
                  "in": "header",
                  "name": "Authorization",
                  "schema": {
                     "type": "string",
                     "format": "JWT"
                  },
                  "required": true
               }
            ],
            "responses": {
               "200": {
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Status"
                        }
                     }
                  },
                  "description": "OK"
               },
               "401": {
                  "description": "Unauthorized access"
               },
               "404": {
                  "description": "Status not found"
               }
            },
            "security": [
               {
                  "bearerAuth": []
               }
            ]
         }
      },
      "/status/{statusId}": {
         "get": {
            "tags": [
               "status"
            ],
            "summary": "Get detail movie status data in particular statusId",
            "operationId": "getMovieStatusByID",
            "parameters": [
               {
                  "name": "statusId",
                  "in": "path",
                  "description": "ID of movie status to retrieve specific status data",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "format": "int32"
                  }
               },
               {
                  "in": "header",
                  "name": "Authorization",
                  "description": "Authorization bearer token",
                  "required": true,
                  "schema": {
                     "type": "string",
                     "format": "JWT"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "successful operation",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Status"
                        }
                     }
                  }
               },
               "400": {
                  "description": "Invalid ID supplied"
               },
               "401": {
                  "description": "Unauthorized access"
               },
               "404": {
                  "description": "Level not found"
               }
            },
            "security": [
               {
                  "bearerAuth": []
               }
            ]
         },
         "put": {
            "tags": [
               "status"
            ],
            "summary": "Update specific movie status data according to status ID",
            "operationId": "updateMovieStatusByID",
            "parameters": [
               {
                  "name": "statusId",
                  "in": "path",
                  "description": "ID of movie status to retrieve specific status data",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "format": "int32"
                  }
               },
               {
                  "name": "Authorization",
                  "in": "header",
                  "required": true,
                  "schema": {
                     "type": "string",
                     "format": "JWT"
                  }
               }
            ],
            "requestBody": {
               "description": "Updated movie status object",
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "anyOf": [
                           {
                              "$ref": "#/components/schemas/Status"
                           }
                        ]
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "successful operation",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Status"
                        }
                     }
                  }
               },
               "400": {
                  "description": "Invalid ID supplied"
               },
               "401": {
                  "description": "Unauthorized access"
               },
               "404": {
                  "description": "Status not found"
               }
            },
            "security": [
               {
                  "bearerAuth": []
               }
            ]
         },
         "delete": {
            "tags": [
               "status"
            ],
            "summary": "Delete movie status data based on defined ID",
            "description": "",
            "operationId": "deleteMovieStatus",
            "parameters": [
               {
                  "in": "header",
                  "name": "Authorization",
                  "required": true,
                  "schema": {
                     "type": "string",
                     "format": "JWT"
                  }
               },
               {
                  "name": "statusId",
                  "in": "path",
                  "description": "ID of movie status to retrieve specific movie status data",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "format": "int32"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "OK",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "message": {
                                 "type": "string"
                              }
                           }
                        }
                     }
                  }
               },
               "400": {
                  "description": "Invalid ID supplied"
               },
               "401": {
                  "description": "Unauthorized access"
               },
               "404": {
                  "description": "Status not found"
               }
            },
            "security": [
               {
                  "bearerAuth": []
               }
            ]
         }
      },
...
   },
   "components": {
      "securitySchemes": {
         "bearerAuth": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
         }
      },
      "schemas": {
...
         "Genre": {
            "type": "object",
            "description": "Movies genre",
            "properties": {
               "id": {
                  "type": "integer",
                  "format": "int32"
               },
               "name": {
                  "type": "string"
               },
               "created_at": {
                  "type": "string",
                  "format": "date-time"
               },
               "updated_at": {
                  "type": "string",
                  "format": "date-time"
               }
            }
         },
         "Status": {
            "type": "object",
            "properties": {
               "id": {
                  "type": "integer",
                  "format": "int32"
               },
               "name": {
                  "type": "string",
                  "enum": [
                     "rumored",
                     "planned",
                     "in production",
                     "post production",
                     "released",
                     "cancelled"
                  ]
               },
               "created_at": {
                  "type": "string",
                  "format": "date-time"
               },
               "updated_at": {
                  "type": "string",
                  "format": "date-time"
               }
            }
         },
...
      }
   },
...
}